backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.4.1
python-dotenv==1.1.0
requests==2.32.3
matplotlib==3.10.8
seaborn==0.13.2
numpy==1.23.5
google-generativeai==0.8.5
openai==1.82.0
eventlet==0.37.0